requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
cryptography==35.0.0
discord-interactions==0.3.0
discord-py-slash-command==3.0.1
discord.py==1.7.3
h11==0.12.0
idna==3.3
multidict==5.2.0
outcome==1.1.0
pycparser==2.21
PyNaCl==1.4.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
requests==2.26.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.0
urllib3==1.26.7
wsproto==1.0.0
yarl==1.7.2
